+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - fix: initial_gas set to max_fee and fixed fee not being charged when max_fee=0
 - fix: correct value of compiled_class_hash in RPCTransaction
 - ci: scope cache by branch and add cache cleanup
+- test: add `starknet-rpc-test` crate to the workspace
+- test(rpc): add `get_block_number.rs` tests
 
 ## v0.2.0